requirements.txt
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
dnspython==2.4.2
idna==3.4
IPy==1.1
numpy==1.26.1
pandas==2.1.1
pycparser==2.21
pymongo==4.5.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
ripe.atlas.cousteau==2.0.0
ripe.atlas.sagan==1.3.1
ripe.atlas.tools==3.1.0
six==1.16.0
socketIO-client==0.7.2
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
websocket-client==1.3.3
